refactor: clean bookmark icons
