+++ packages/blocks/src/icons/text.ts
@@ -442,6 +442,30 @@ export const SearchIcon = html`
   </svg>
 `;
 
+export const RefreshIcon = html`<svg
+  xmlns="http://www.w3.org/2000/svg"
+  width="20"
+  height="20"
+  viewBox="0 0 24 24"
+>
+  <path
+    fill-rule="evenodd"
+    clip-rule="evenodd"
+    d="M3.25 12C3.25 7.16751 7.16751 3.25 12 3.25C14.2406 3.25 16.286 4.09311 17.8335 5.4782L17.8359 5.48034L19.25 6.75695V4C19.25 3.58579 19.5858 3.25 20 3.25C20.4142 3.25 20.75 3.58579 20.75 4V8.44444C20.75 8.85866 20.4142 9.19444 20 9.19444H15.5556C15.1413 9.19444 14.8056 8.85866 14.8056 8.44444C14.8056 8.03023 15.1413 7.69444 15.5556 7.69444H18.05L16.8331 6.5959C16.8327 6.59554 16.8323 6.59518 16.8319 6.59483C15.5489 5.44701 13.8568 4.75 12 4.75C7.99594 4.75 4.75 7.99594 4.75 12C4.75 16.0041 7.99594 19.25 12 19.25C15.4494 19.25 18.3378 16.8403 19.0704 13.6118C19.162 13.2079 19.5638 12.9547 19.9677 13.0464C20.3717 13.138 20.6248 13.5398 20.5332 13.9437C19.6489 17.8405 16.1651 20.75 12 20.75C7.16751 20.75 3.25 16.8325 3.25 12Z"
+  />
+</svg>`;
+
+export const WebIconSVG = svg`<g id="Web">
+  <path
+    id="Shape (Stroke)"
+    fill-rule="evenodd"
+    clip-rule="evenodd"
+    d="M3.78363 11.25H8.26C8.32021 8.99973 8.64952 6.95181 9.16717 5.39887C9.32825 4.91563 9.51187 4.46715 9.7189 4.06941C6.50351 4.99257 4.09288 7.81832 3.78363 11.25ZM12 2.25C6.61522 2.25 2.25 6.61522 2.25 12C2.25 17.3848 6.61522 21.75 12 21.75C17.3848 21.75 21.75 17.3848 21.75 12C21.75 6.61522 17.3848 2.25 12 2.25ZM12 3.75C11.9141 3.75 11.7129 3.80155 11.4213 4.17149C11.1364 4.53298 10.8471 5.10255 10.5902 5.87321C10.1323 7.24686 9.82057 9.12821 9.76056 11.25H14.2394C14.1794 9.12821 13.8677 7.24686 13.4098 5.87321C13.1529 5.10255 12.8636 4.53298 12.5787 4.17149C12.2871 3.80155 12.0859 3.75 12 3.75ZM15.74 11.25C15.6798 8.99973 15.3505 6.95181 14.8328 5.39887C14.6718 4.91563 14.4881 4.46715 14.2811 4.06941C17.4965 4.99257 19.9071 7.81832 20.2164 11.25H15.74ZM14.2394 12.75H9.76056C9.82057 14.8718 10.1323 16.7531 10.5902 18.1268C10.8471 18.8975 11.1364 19.467 11.4213 19.8285C11.7129 20.1985 11.9141 20.25 12 20.25C12.0859 20.25 12.2871 20.1985 12.5787 19.8285C12.8636 19.467 13.1529 18.8975 13.4098 18.1268C13.8677 16.7531 14.1794 14.8718 14.2394 12.75ZM14.2811 19.9306C14.4881 19.5329 14.6718 19.0844 14.8328 18.6011C15.3505 17.0482 15.6798 15.0003 15.74 12.75H20.2164C19.9071 16.1817 17.4965 19.0074 14.2811 19.9306ZM9.7189 19.9306C9.51187 19.5329 9.32825 19.0844 9.16717 18.6011C8.64952 17.0482 8.32021 15.0003 8.26 12.75H3.78363C4.09288 16.1817 6.50351 19.0074 9.7189 19.9306Z"
+  />
+</g>`;
+
+export const WebIcon16 = icon(WebIconSVG, 16);
+
 // Link Icon
 
 /**
